+++ app/src/main/java/dev/nelson/mot/callback/SetDataFromPaymentLoaderCallbacks.java
@@ -2,5 +2,5 @@ package dev.nelson.mot.callback;
 
 
 public interface SetDataFromPaymentLoaderCallbacks {
-    public void fillPaymentInitialStateWithData(String title, int categoryId, String categoryName, double cost, String summary);
+    public void fillPaymentInitialStateWithData(String title, int categoryId, String categoryName, long cost, String summary);
 }
